Add support for eprint and doi information to the references
Add support for two extra bibtex fields. These are eprint and doi.
Each of these can be specified as a field in the bibtex, and will cause
the references to contain eg: arXiv:1209.1234 and doi:10.0.0.1/XXXX.

The style of each of these is controlled by the \eprint and \doi
commands.

Because they are specified as commands, the user can over-ride them to
be hyperlinks in the preamble to their tex file. The intro to the bst
file shows how this may be done.

Builds on an idea originally by Antony Lewis antony@cosmocoffee.info