Add ident attribute
We would like to make sure that we can work out what version people
are downloading without necessarily having to put in explicit version
numbers. We do attempt to update the HISTORY field when we make changes
but in addition we now use a ident git attribute to automatically
embed the SHA1 in the file when it is checked out.

This fixes issue #3