Add DOIs and eprint links to non-articles. Relevant to issue #6, but does not, I think, close it.
We include a format.doi, format.eprint and format.url in all entry types
_except_ {article}.  These will be redundant in most cases: at present (2013)
the majority of things that have DOIs, or which are submitted to arXiv, are
articles; however it's not impossible for other things to have the extra links,
so we keep the printed links for the sake of consistency.

In articles, any DOI is linked to the journal name, and any 'adsurl' field
(which will generally be present on BibTeX entries downloaded from ADS) will be
linked to the journal and page number (suggestion from Keith).  Note that we do
this _only_ if there is an 'adsurl' field present; it's tempting, but surely too
errorprone, to start synthesising ADS URLs from scratch.