Handle presence/absence of eprint prefix.
Addresses issue #7, but I'd like some discussion before closing it.

This makes minimal changes to the .bst file, but replaces \eprint by a
two-argument version, and \doi by a version with an optional first argument.
These two macros are removed from the .bbl output, and replaced with same-name
ones in the preamble of mn2e-example.tex; these should be considered for
migration to mn2e.cls.

These appear to work, but they do mean that the mn2e.bst output would be usable
_only_ with mn2e.cls.

I've implemented support for archivePrefix="dblp", not because I think very many
people will want to use a CS bibliography archive, but to illustrate how support
for other archives is provided.  Ie, it's not 'arXiv or bust'.