Fix bibliography sort order
The goal here is to be compatible with the MN style prescription that:
    If there are several references with the same first author,
    arrange in the following order:
    firstly single-author papers (by date);
    then two-author papers (alphabetically by co-author, then by date);
    then multi-author papers (by date).
We deem 'Jones and others' to be a multi-author paper, in this sense,
so that it sorts, by date, with the three-or-more author papers.
An alternative is to deem 'Jones and others' to sort just after the single-author papers
(this case isn't really covered by the above rules, but this position looks sensible).