Delete function output.year.misc.check which appeared to be both erroneous and redundant (not good).
This also deleted the rather puzzled comment before the sole use of the
function, in {misc}, which I reproduce below for posterity.

The following at one time included output.year.misc.check rather
than output.year.check.  It's not clear why, since this seems to
have inevitably resulted in an extra comma in this case (ie,
Jones A, Smith B., , 2000, ....").  Use plain output.year.check
until it becomes clear what the function of the (erroneous)
output.year.misc.check is or was.