Enable arXiv identifier printing if bib entry is from arXiv
Enabled ArXiv e-print identifier to be printed for articles that
are pre-prints only (and have "ArXiv e-print" in the journal field).
Still some problems with the use of commas.  Only 1 comma will appear
between the year and "arXiv..." for such preprints, but for e.g.
conference precedings there may be no comma at all